+++ temporalio/nexus/__init__.py
@@ -1,11 +1,3 @@
-import logging
-from typing import (
-    Any,
-    Mapping,
-    MutableMapping,
-    Optional,
-)
-
 from ._decorators import workflow_run_operation as workflow_run_operation
 from ._operation_context import Info as Info
 from ._operation_context import (
@@ -19,28 +11,10 @@ from ._operation_context import (
 )
 from ._operation_context import client as client
 from ._operation_context import info as info
+from ._operation_context import logger as logger
 from ._operation_handlers import cancel_operation as cancel_operation
 from ._token import WorkflowHandle as WorkflowHandle
 
-
-class LoggerAdapter(logging.LoggerAdapter):
-    def __init__(self, logger: logging.Logger, extra: Optional[Mapping[str, Any]]):
-        super().__init__(logger, extra or {})
-
-    def process(
-        self, msg: Any, kwargs: MutableMapping[str, Any]
-    ) -> tuple[Any, MutableMapping[str, Any]]:
-        extra = dict(self.extra or {})
-        if tctx := _temporal_operation_context.get(None):
-            extra["service"] = tctx.nexus_operation_context.service
-            extra["operation"] = tctx.nexus_operation_context.operation
-            extra["task_queue"] = tctx.info().task_queue
-        kwargs["extra"] = extra | kwargs.get("extra", {})
-        return msg, kwargs
-
-
-logger = LoggerAdapter(logging.getLogger(__name__), None)
-"""Logger that emits additional data describing the current Nexus operation."""
 
 # TODO(nexus-prerelease) WARN temporal_sdk_core::worker::nexus: Failed to parse nexus timeout header value '9.155416ms'
 # 2025-06-25T12:58:05.749589Z  WARN temporal_sdk_core::worker::nexus: Failed to parse nexus timeout header value '9.155416ms'
